requirements.txt
yara-python
pefile
requests
scikit-learn